# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: metarang
    description: REST API for metarang microservices. Legacy-compatible JSON responses served via Kong → service HTTP APIs.
    title: metarang API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: support-service HTTP direct (dev)
      url: http://localhost:8070
...
paths:
...
    /api/hourly-profits:
        get:
            operationId: get_hourly-profits
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List hourly profits or filter by application
            tags:
                - Features
        post:
            operationId: post_hourly-profits
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
                required: true
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List hourly profits or filter by application
            tags:
                - Features
...
    /api/maps:
        get:
            operationId: get_maps
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List maps
            tags:
                - Maps
...
components:
    schemas:
...
        Error:
            properties:
                message:
                    type: string
            type: object
...
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http